level `int` -> `byte`
Might refactor the learnset level get to byte later as a TryGet so -1 is never returned.

+++ PKHeX.Core/Editing/PKM/EntitySuggestionUtil.cs
@@ -9,7 +9,7 @@ namespace PKHeX.Core;
 /// </summary>
 public static class EntitySuggestionUtil
 {
-    public static List<string> GetMetLocationSuggestionMessage(PKM pk, int level, ushort location, int minimumLevel, IEncounterable? enc)
+    public static List<string> GetMetLocationSuggestionMessage(PKM pk, byte level, ushort location, int minimumLevel, IEncounterable? enc)
     {
         var suggestion = new List<string> { MsgPKMSuggestionStart };
         if (pk.Format >= 3)
