+++ PKHeX.Core/Legality/Encounters/Templates/GO/EncounterSlot7GO.cs
@@ -107,7 +107,7 @@ public sealed record EncounterSlot7GO(int StartDate, int EndDate, ushort Species
         }
     }
 
-    private void SetEncounterMoves(PB7 pk, int level)
+    private void SetEncounterMoves(PB7 pk, byte level)
     {
         Span<ushort> moves = stackalloc ushort[4];
         var source = LearnSource7GG.Instance;
